crates/sel4-microkit/smoltcp-device-adapter: Split into modules
